Removing the check for issue 64. We already blow up when trying to make Class or Constructor's members accessible, via the new private-constructor check.
I attempted to add a test to see whether Guice provides reasonable errors if a SecurityManager is present, but it seems like it's not really worth the effort. Almost everything we do with java.lang.Reflect is going to need reflect access, and handling this gracefully would require wrapping EVERYTHING in try/catch blocks for SecurityException.  We may consider adding a check at the beginning of say, Guice.createInjector to ensure that reflection is available, but I don't think it's critical for the time being.

Also added some tests for requestInjection to make sure errors have the appropriate source lines. We weren't including source lines properly for members added via requestInjection. This is fixed.