Adds Graphviz-based grapher extension
Adds grapher extension, an implementation for a Graphviz-based tool to
visualize an Injector. Main class is InjectorGrapher, which is available from
the GrapherModule. The GraphvizModule should also be installed for Graphviz
output.

See com.google.inject.grapher.demo.InjectorGrapherDemo for an example and to
generate a sample graph.

Fix for issue 213.