MiniGuice, a proof of concept.
The standard Guice injector hits a diverse set of use cases with a small core. It has advanced features to make developers productive and it's very fast. And it scales to extremely large teams very well. In a world of JavaEE aircraft carriers, Guice is a sports car. 

To contrast, MiniGuice (placeholder name) is roller skates. It has barely any features and can't do anything but the bare essentials: field & constructor injection, provider methods, provider injections, singletons, and binding annotations. It's intended for use in tiny dozen-file projects maintained by a single developer in leisure time. It's only features are its compact size and implementation simplicity.